+++ services/sms/include/sms_service.h
@@ -36,7 +36,7 @@ public:
     int32_t Dump(std::int32_t fd, const std::vector<std::u16string> &args) override;
     std::string GetBindTime();
     void InsertSessionAndDetail(int32_t slotId, const std::string &telephone, const std::string &text,
-        uint16_t &dataBaseId);
+        int32_t &dataBaseId);
 
     /**
      * Sends a text Type SMS message.
@@ -341,7 +341,7 @@ public:
      * @param time Indicates tiemstamp of hap call NAPI function
      * @return Returns id of mms in sqlite
      */
-    uint16_t QueryMmsDatabaseId(int32_t slotId, int64_t time);
+    int32_t QueryMmsDatabaseId(int32_t slotId, int64_t time);
 
     /**
      * Service after Send Mms.
@@ -354,7 +354,7 @@ public:
      * @param sendStatus status of mms sent
      * @return Returns {@code 0} if send mms success; returns {@code false} otherwise
      */
-    void ServiceAfterSendMmsComplete(int32_t slotId, int64_t &time, uint16_t &dataBaseId,
+    void ServiceAfterSendMmsComplete(int32_t slotId, int64_t &time, int32_t &dataBaseId,
         DataShare::DataShareValuesBucket &sessionBucket, std::string  &sendStatus);
 
     /**
@@ -380,7 +380,7 @@ private:
     void TrimSmscAddr(std::string &sca);
     bool CheckSimMessageIndexValid(int32_t slotId, uint32_t msgIndex);
     void InsertSmsMmsInfo(int32_t slotId, uint16_t sessionId, const std::string &number,
-        const std::string &text, uint16_t &dataBaseId);
+        const std::string &text, int32_t &dataBaseId);
     bool InsertSession(bool isNewSession, uint16_t messageCount, const std::string &number, const std::string &text);
     bool QuerySessionByTelephone(const std::string &telephone, uint16_t &sessionId, uint16_t &messageCount);
     void UpdateSmsContact(const std::string &address);
